Implement proper multiplicative zoom with float step_size
- Changed step_size from u32 to f64 throughout the codebase
- Implemented correct multiplicative zoom tracking:
  * Track aggregate zoom factor in WindowState
  * Issue multiply commands incrementally using step_size
  * Calculate final partial multiply to reach exact target without overshooting
  * Example: step=2x, target=9x issues 2x,2x,2x,(9/8)x
- Updated KittyRegistry.multiply_font_size_by and divide_font_size_by to accept f64
- Updated execute_font_command_with_op to handle multiplicative ops (issue single command)
- Updated config.example.toml with correct multiplicative zoom documentation
- Fixed all type errors and warnings

+++ src/config.rs
@@ -79,8 +79,8 @@ fn default_reap_interval() -> u64 {
     300 // 5 minutes
 }
 
-fn default_step_size() -> u32 {
-    1
+fn default_step_size() -> f64 {
+    1.0
 }
 
 /// Zoom type: absolute, additive, or multiplicative
@@ -110,7 +110,7 @@ pub struct ZoomConfig {
 
     /// Number of steps to apply at once
     #[serde(default = "default_step_size")]
-    pub step_size: u32,
+    pub step_size: f64,
 }
 
 impl Default for ZoomConfig {
@@ -191,7 +191,7 @@ pub struct CliZoomArgs {
     pub absolute: Option<f64>,
     pub additive: Option<f64>,
     pub multiplicative: Option<f64>,
-    pub step_size: Option<u32>,
+    pub step_size: Option<f64>,
 }
 
 #[derive(Debug, Clone, Serialize, Deserialize)]
